requirements.txt
aiohttp==3.7.4.post0
aiojobs==0.3.0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
google-auth==1.27.1
idna==2.10
iso8601==0.1.14
kopf==1.30.1
kubernetes==12.0.1
multidict==5.1.0
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-json-logger==2.0.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.58.0
yarl==1.6.3
